+++ scm-webapp/src/test/java/sonia/scm/api/v2/resources/ChangesetRootResourceTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.api.v2.resources;
 
 
@@ -43,6 +43,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import sonia.scm.repository.Changeset;
 import sonia.scm.repository.ChangesetPagingResult;
+import sonia.scm.repository.ChangesetTrailers;
 import sonia.scm.repository.NamespaceAndName;
 import sonia.scm.repository.Person;
 import sonia.scm.repository.Repository;
@@ -56,6 +57,7 @@ import java.net.URI;
 import java.time.Instant;
 import java.util.Date;
 import java.util.List;
+import java.util.Set;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -67,7 +69,6 @@ import static org.mockito.Mockito.when;
 @RunWith(MockitoJUnitRunner.Silent.class)
 @Slf4j
 public class ChangesetRootResourceTest extends RepositoryTestBase {
-
 
   public static final String CHANGESET_PATH = "space/repo/changesets/";
   public static final String CHANGESET_URL = "/" + RepositoryRootResource.REPOSITORIES_PATH_V2 + CHANGESET_PATH;
@@ -86,6 +87,10 @@ public class ChangesetRootResourceTest extends RepositoryTestBase {
   @Mock
   private LogCommandBuilder logCommandBuilder;
 
+  @Mock
+  private Set<ChangesetTrailers> changesetTrailers;
+
+  @InjectMocks
   private ChangesetCollectionToDtoMapper changesetCollectionToDtoMapper;
 
   @InjectMocks
@@ -93,10 +98,8 @@ public class ChangesetRootResourceTest extends RepositoryTestBase {
 
   private ChangesetRootResource changesetRootResource;
 
-
   private final Subject subject = mock(Subject.class);
   private final ThreadState subjectThreadState = new SubjectThreadState(subject);
-
 
   @Before
   public void prepareEnvironment() {